Mario's Jupyter notebooks
I added two notebooks.
First is Testing_AudioAnalysis -- this is the one I showed this morning.
I import a couple of audio files, look at the output of pyAudioAnalysis,
and play around with the length of the result. I also produce a Data
Frame with the actors, sentence, emotion, and intensity of each clip.
The other notebook, Spectrograms, produces the Mel Spectrograms.