Remove blank audio file
The file Data/Train/1076_MTI_SAD_XX.wav is empty. It's silent when I try
to play it and all its feature vectors are empty. For now, I've removed
the file and its entries from Categories.csv and Categories_train.csv.
Maybe I'll look if this mistake is in any other file, but since we
aren't using the midFeatures directly, this probably won't bother.
**IMPORTANT!** When modifying dataframes, make sure that you're adding
data to the correct rows!