Removed .DS_Store and added .gitignore_global
.DS_Store is a file automatically generated by Mac computers. It should
not be in a repo because every Mac user has their own .DS_Store and so
this introduces unnecessary conflicts.
To solve this, I removed all .DS_Store files in the repo and subfolders
and added a .gitignore_global file to prevent it from being added again.
Mac users might want to check
https://stackoverflow.com/questions/18393498/gitignore-all-the-ds-store-files-in-every-folder-and-subfolder